Discard intervals that we aren't interested in.
Based on hints provided on an interval.
We check if the intervals are top level (i.e. will get values printed)
Or if they can be snipped during shrinking,
Or if they are required for uniqueness.
If none of those are true then we don't need them, and they can be safely discarded.