Sort out weird rebinding issue with test.check.
Rebinding the report function into a plain fn was messing with test.check.
During the load of test.check.clojure-test, it tries to attach a method
to the report fn. This fails since I've already rebound it to a plain fn.
It doesn't seem to cause any issues since I don't use the associated
functionality, but it seems to be pulled in as part of spec checking.
Since that's late bound it explodes during the first test that's run.