Use buffers inside sources.
Seems like there might be some kind of non-deterministic issue going on.
My suspicion is it's due to the interplay between atoms and the java
class. Since I'm going to change the way the chained buffer is used,
which should allow me to pull it off the atom, I'm going to wait until
that's done before I try and debug this some more.